+++ src/components/camera/index.tsx
@@ -5,12 +5,9 @@ import {
 } from "@react-three/drei";
 import { useFrame, useThree } from "@react-three/fiber";
 import * as THREE from "three";
-import { useEffect } from "react";
 import "./camera.css";
 import cameraPos from "../../utils/cameraPos.json";
-import gsap from "gsap";
-import anim from "./kframe.json";
-const kframe = anim.frames;
+import animate from "./animate";
 
 // Disable react/prop-types for this file
 /* eslint-disable react/prop-types */
@@ -19,48 +16,6 @@ interface CameraProps {
   helper?: boolean; // Optional boolean prop to control the helper display
   fly?: boolean; // Optional boolean prop to control the fly controls
 }
-
-const animate = (
-  camera: THREE.PerspectiveCamera | THREE.OrthographicCamera,
-) => {
-  useEffect(() => {
-    const timeline = gsap.timeline({ repeat: 0, repeatDelay: 1 });
-
-    kframe.forEach((frame, index) => {
-      // Animate camera position
-      timeline.to(camera.position, {
-        x: frame.position.x,
-        y: frame.position.y,
-        z: frame.position.z,
-        duration: 2,
-        ease: "power2.out",
-        onUpdate: () => camera.updateProjectionMatrix(),
-      });
-
-      // Animate camera lookAt
-      timeline.to(
-        {},
-        {
-          duration: 2,
-          onUpdate: () => {
-            const lookAt = new THREE.Vector3(
-              frame.lookingAt.x,
-              frame.lookingAt.y,
-              frame.lookingAt.z,
-            );
-            camera.lookAt(lookAt);
-          },
-        },
-        index * 2, // Sync lookAt with position animation
-      );
-    });
-
-    // Return a cleanup function that stops the timeline
-    return () => {
-      timeline.kill();
-    };
-  }, [camera]);
-};
 
 const Camera: React.FC<CameraProps> = ({ helper = false, fly = false }) => {
   const { camera } = useThree();
